+++ src/onegov.ballot/onegov/ballot/models/common.py
@@ -1,4 +1,6 @@
+from onegov.core.orm.types import JSON
 from sqlalchemy import Column, Enum
+from sqlalchemy.ext.declarative import declared_attr
 
 
 class DomainOfInfluenceMixin(object):
@@ -13,3 +15,16 @@ class DomainOfInfluenceMixin(object):
         ),
         nullable=False
     )
+
+
+class MetaMixin(object):
+    """ Mixin providing a meta/content JSON pair. Meta is a JSON column loaded
+    with each request, content is a JSON column loaded deferred (to be shown
+    only in the detail view).
+
+    """
+
+    #: metadata associated with the form, for storing small amounts of data
+    @declared_attr
+    def meta(cls):
+        return Column(JSON, nullable=False, default=dict)