+++ src/onegov.ballot/onegov/ballot/models/election.py
@@ -1,5 +1,5 @@
 from collections import OrderedDict
-from onegov.ballot.models.common import DomainOfInfluenceMixin
+from onegov.ballot.models.common import DomainOfInfluenceMixin, MetaMixin
 from onegov.core.orm import Base, translation_hybrid
 from onegov.core.orm.mixins import TimestampMixin
 from onegov.core.orm.types import HSTORE, UUID
@@ -32,7 +32,7 @@ class DerivedBallotsCount(object):
 
 
 class Election(Base, TimestampMixin, DerivedBallotsCount,
-               DomainOfInfluenceMixin):
+               DomainOfInfluenceMixin, MetaMixin):
 
     __tablename__ = 'elections'